getPayout moved into a separate file, add tests
