+++ services/sentry/lib/eventReducer.js
@@ -1,5 +1,6 @@
 const BN = require('bn.js')
 const toBalancesKey = require('../../toBalancesKey')
+const getPayout = require('./getPayout')
 
 function newAggr(channelId) {
 	return { channelId, created: new Date(), events: {}, totals: {}, earners: [] }
@@ -31,20 +32,6 @@ function reduce(channel, initialAggr, ev) {
 	}
 
 	return aggr
-}
-
-function getPayout(channel, ev) {
-	if (ev.type === 'IMPRESSION' && ev.publisher) {
-		// add the minimum price for the event to the current amount
-		return [toBalancesKey(ev.publisher), new BN(channel.spec.minPerImpression || 1)]
-	}
-	if (ev.type === 'CLICK' && ev.publisher) {
-		return [
-			toBalancesKey(ev.publisher),
-			new BN((channel.spec.pricingBounds && channel.spec.pricingBounds.CLICK.min) || 0)
-		]
-	}
-	return null
 }
 
 function mergeEv(initialMap = { eventCounts: {}, eventPayouts: {} }, [earner, amount]) {
